fix: move function out of default module export and update test.
